Use type-appropriate math functions

+++ src/gui/rgba.hpp
@@ -34,7 +34,7 @@ rgba_to_uint(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 static inline uint8_t
 mono_interpolate(uint8_t v1, uint8_t v2, float f)
 {
-	return static_cast<int>(rint((v2) * (f) + (v1) * (1 - (f))));
+	return static_cast<int>(rintf((v2) * (f) + (v1) * (1 - (f))));
 }
 
 #define RGBA_R(x) (static_cast<uint32_t>(x) >> 24)
